(PUP-9071) Fix portage package provider to support specific versioning
In some cases you want to install a specific package version like "1.2.3-r4"
of the package "example/package". If you want to install it from the
commandline you would issue the command:

emerge =example/package-1.2.3-r4

Unfortunately the portage package provider does issue the following command
which is missing the version informations.

emerge example/package

As a result, on every puppet run it wants to have the version "1.2.3-r4" to
be installed, and tries to install the desired package version, but never
succeeds with the correct version.

To achieve the desired functioning of the portage package provider, it was
necessary to update the query commands for the helper programms qatom and eix,
to have the correct emerge commandline arguments, when installing a specific
version of a package.

The missing tests were also added for the installation of specific versions
and slots of packages.