(maint) Update puppetlabs-puppetserver_perf_control submodule
The puppetlabs-puppetserver_perf_control is used during the benchmark
tests, and assumes that certain types are available. Some of the types
that it assumed were available have been removed from the Puppet
repository, as they have been extracted into modules of their own. Because
of this, the puppetlabs-puppetserver_perf_control repo has been updated to
pull in these modules via its Puppetfile.