(PUP-9077) Fix problems with 3.x functions returning :undef
Since the calling API for 3.x functions requires nil values to be
transformed to Ruby Symbol :undef such values could be returned from 3.x
functions thus causing 4.x functions to also get such values.

This is bad because :undef is a 3.x construct that should not be
exposed in Puppet's modern APIs.

This commit fixes this by rewriting all returned values from 3.x
functions - transforming any nested :undef found in Array or Hash
values. The rewrite does not replace :undef in other types of values.

In this commit, the ToDataConverter is also updated to handle
an :undef value in the rare event that it escaped from a 3.x runtime.