(PUP-6919) Add :cwd option to Puppet::Util::Execution.execute
This will run the command in the specified working directory. Not only
is it a useful option to have, we also need this option for Windows to
implement PUP-6920. It also paves the way for the work in PUP-9150.

Note that this option is only available on the agent. It cannot be used
on the master.