(PUP-6919) Add :cwd option to Puppet::Util::Execution.execute
This will run the command in the specified working directory. We need
this option for Windows specifically since the :cwd could be specified in
different pathing formats that may not work correctly with Ruby's chdir method
on some platforms (e.g. like UNC paths on Windows machines).

Note that this option is only available on the agent. It cannot be used
on the master.