(PUP-6964) Make 4.x loaders load 3.x functions
This adds loading of 3.x functions using the 4.x loaders.
As a result, all 3.x functions will be loaded with the lower
precedence than 4.x functions, but now computed per module,
and they can be called exactly the same way as 4.x functions
as well as by using the 3.x calling convention.

3.x functions that are not autoloaded (manually added, primarily in
tests) are called only if autoaloding did not find an override. This is
compatible with the old precedence.