(PUP-6964) Make the 4x wrapped 3x func bypass dispatcher in call
This makes the call to a 3x function via a 3x wrapper short circuit
the dispatcher when making a call. This is done because raised error
for mismatching argument count would otherwise not be compatible.
It was simply too complicated to make the 4.x ArgumentMismatchDescriber
compatible with both 4.x and 3.x expectations in how errors are
reported.