(PUP-6964) Refactor where return from 3x func is converted
There was a problem with 3x function return conversion
as it needs to be done in multiple places since call paths
are different. This moves the mapping of the return to
the 3x function "dispatcher" (the function_xxx method). Thus making
sure all calls to 3x get the same treatment irrespective of
from where it is called.

Two tests are changed as they assumed that 3x would return
unfiltered values (containing `:undef`).