(PUP-7433) Add optional features
Include optional gems in the "features" group to increase test coverage:

    diff
    ldap
    minitar
    ssh
    telnet

The "ldap" feature (and related Puppet::Util::Ldap code) is based on the
old ruby-ldap gem, not the newer net-ldap gem.

These features are always present in modern ruby versions, so no
additional gems are needed to enable them.

    openssl
    posix
    syslog
    zlib

These features are never tested in puppet, but are left as-is to not
break other projects, eg bolt.

    bolt
    sqlite
    pe_license

These features are excluded because they rely on highly platform
dependent gems, and are best handled by acceptance tests:

    cfpropertylist
    eventlog
    libuser
    libshadow
    selinux