(maint) Remove remaining references to CA settings
This commit cleans up remaining references to CA settings. We have opted
not to remove their definitions from the defaults file for Puppet 6,
and they will continue to be used by Puppet Server, but there is no need
for Puppet's code to continue to reference them now that the CA code is
gone.