(maint) Update ldap tests so they work with mocha
The transpec migration from rspec 2 to 3 in
125712f7f0ef60b8edae21702ab4617c2e893aea changed usages of `stub` and
`mock` to `double`, but we're using mocha not rspec-mocks. This renaming
was fixed in most places except the ldap terminus, because the tests were
always skipped (since the ruby-ldap gem wasn't installed).

Now that we are installing the gem, update the tests so they work with
mocha.