(maint) Guard tests with ldap feature
The manager unit test defined an LDAP class that conflicted with the
LDAP module that ruby-ldap defines. There are also problems with the
test exception hierarchy. Avoid all of that by guarding the manager
tests with the ldap feature which is now installed by default on MRI
ruby platforms (not jruby or windows).