(PUP-9092) Improve process termination handling on Windows
Previously we waited a minimum of 1 second in ruby for the child process
to exit, which introduced an unnecessary delay for short lived
processes.

Now we wait at most 15 milliseconds, and if we timeout, sleep for 0
seconds in ruby, and repeat. This provides better resolution if the
process exits quickly.