(PUP-6919) Change to the cwd before changing privileges in execute_posix
Previously in execute_posix, we'd cd into the cwd after changing the
process' privileges to run under a different user. This was causing
failures for the case when the user did not have permissions to access
the cwd.

This commit changes to the cwd before changing privileges.

This commit adds an acceptance test that replicates this bug, and also adds another
acceptance test to ensure that the exec resource can still run the command under a
different user even if that user does not have permissions to access the
pwd. The latter acceptance test is how this bug was originally detected
and is also a separate requirement of the exec resource.