(PUP-8967) Request simple status to detect if server is functional
Previously, puppet would make a node request to determine which server
in its `server_list` was functional. The node request is non-trivial as
puppetserver needs to acquire a JRuby instance, query puppetdb and
classifier for node information, and return it. The agent normally
passed the node object through to the normal agent run, so the effort
wasn't wasted. But if `strict_environment_mode` is enabled, then the
node was never used.

Also, if there is a classification error, then the agent will interpret
that to mean the server is unreachable until it exhausts all servers in
the list.

This commit switches the agent to sending a HEAD request to the
/status/v1/simple endpoint. This means the agent will make at least one
additional HTTP request per run when using `server_list`. But it should
be negligible because the connection is persisted, puppetserver handles
the request in clojure, and doesn't query puppetserver metrics. In fact
the HEAD response body will be empty.