(PUP-9028) Create ssl application for managing client keys/certs
Create an ssl application for managing client keys and certs. Currently
it can only generate and submit a certificate request. As part of the
bootstrapping process, the application will download the CA and CRL
bundles if they don't exist, and use them for subsequent REST requests.
The application will also generate a key pair if it doesn't exist, and
use that to generate the CSR.