(PUP-9093) Add a ToStringifiedConverter
This adds a ToStringifiedConverter that converts any rich data
into Data where all rich data is transformed to a (lossy)
string representation of the value.

This converter is quite permissive in what it allows as values
and does its best to turn something into a human readable string.

It will fail if the value has circular structure (and the circularity
is not in a data type).
(PUP-9093) Improve handling of bad encoding for non ASCII-8BIT

Since any string in non ASCII-8BIT encoding could have illegal byte
sequences, the stringifier now transform all such strings to
UTF-8 (even if already UTF-8) with the option to replace all
undefined and illegal characters to the Unicode 0xFFFD character
(black diamond question mark) defined for this purpose.

(PUP-9093) Make alien data type fact.to_s use UTF-8 encoding if possible

Before this, if a fact value was of an to facts alien data type
it would call fact.to_s to get that value. Depending on the value's
data type, it may call a Kernel method to get that string via
`inspect` and that will unders some circumstances return an
ASCII-8BIT string even if the content is not binary in nature.

This commit changes this so that an encode for UTF-8 takes place and if
that succeeds, then the UTF-8 string content is returned instead of the
binary ASCII-8BIT.

This makes a difference in stringification as all ASCII-8BIT strings
are considered binary and result in Base64 encoded output. This is not
wanted for a string that is non binary in nature.
(PUP-9093) Make the Event class use ToStringifiedConverter

Before this, the Event class produced rich data in its
to_data_hash method. That is not good since the downstream
expectations is that no rich data is present.

This is now changed by using the ToStringifiedConverter replacing
all rich data with suitable string representation for such values.
(This is known to be lossy, and this is desired to be backwards
compatible).

As a consequence, the test for roundtripping a report needed to
change as per the nature of the change, rich data cannot roundtrip.

The Event class itself may use rich data that gets converted, but
it cannot assume it will get them back after sending the report. Tests
needed updating for that.

This commit also adds testing of all conversion done by the
ToStringifiedConverter.