(PUP-9093) Improve error messages involving an Object Type
An error message involving an Object Type would be described as
"Anonymous Class" by the label provider. While that is true, there is
a pcore defined name that can be shown.

This commit changes the label provider to output the actual name
of a pcore defined data type instead of "Anonymous Class".