(PUP-9081) Find required files for a task
To enable including shared files with a task we need to be able to find specified files outside of the tasks own module. This adds a function to find files specified by the task within the modulepath, and concatenates them to the return value of 'files' along with other task files.