(PUP-8967) Request status using GET and require 200 response
The puppetserver simple status endpoint does not accept HEAD requests
and will reply with 404. Update the configurer to request status using
GET and require an HTTP 200 response. Also log when the server is
unavailable, as opposed to unreachable. The simple status response
returns "running" in the body of the message, so the impact should be
negligible.