(PUP-9107) Deprecate ability to change catalog ordering
The ability to set a catalog ordering other than "manifest" is going to be
removed in 6.0.0 (PUP-6165). We now issue a deprecation warning if someone
is using a value other than "manifest", and have updated the setting's
documentation to mention that in 6.0.0 and up it will effectively always
be set to "manifest".