(PUP-9028) Add download_cert action
Add an action for downloading a cert. Verify that the downloaded cert's
public key matches our private key before attempting to save the cert.
If they match, then save the cert, including overwriting an existing
one. This can happen if the cert is cleaned on the server, a new CSR
is submitted using the same keys, and the cert is later issued.