(PUP-9028) Submit the CSR even if we already have a cached version
Previously, if a local CSR existed, then running `submit_request` would
do nothing. This commit adds a `submit_request` method that generates
keys if needed, loads the local CSR if one exists, verifies the CSR
against the local key, and submits the CSR.

We also no longer attempt to download a previously sent CSR, which
eliminates the "puppet never resends a CSR" problem.

We also preserve the behavior of attempting to download the autosigned
cert after submitting the CSR. This is handled as an explicit call to
`download_cert` instead of relying on Host#certificate side-effects.