(PUP-6165) Remove ability to use catalog orderings other than "manifest"
The "manifest" catalog ordering has been the default for a while, and
setting the catalog ordering to anything other than "manifest" was
deprecated in 5.5.x.

As the ticket points out, having the catalog ordering being configurable
means that public modules can't rely on any specific ordering. While this
isn't necessarily a problem if they've fully specified any ordering
dependencies between the resources, it can be easy to miss a dependency.

This removes, the deprecated setting to choose a catalog ordering,
effectively making it permanently set to "manifest".