(PUP-9118) Add context flag for stringifying Resource parameters
Before this, the to_data_hash for a Resource would either return
generic data representing rich data, or a stringified version +
warnings or error (depending on strictness).

For use cases such as PDB that cannot handle rich values a more DB
friendly format is required. In order to allow PDB terminus to get a
data hash of a Resource with suitably stringified parameter values (as
produced by the ToStringifiedConverter) it is now possible to get this
by setting :stringify_rich to true in the context.

```
Puppet.override(:stringify_rich => true, 'example') do
  a_resource.to_data_hash
end
```