(PUP-9079) Make it illegal to use __pvalue and __ptype as attr names
Before this it was possible to create an Object data types with
attribute names __ptype and __pvalue, but when using them they
override the internal values used when serializing.

While there may be a better solution (that allows the names) fixing
that would probably require updating all serialization formats. The
approach in this commit (as discussed on the ticket), is simply to make
it illegal to use the two reserved names __ptype and __pvalue.