(PUP-9110) Make DeferredResolver also resolve Binary as ASCII-8BIT
Before this, Binary values in a catalog resolved to instances
of Puppet::Pops::Types::PBinaryType::Binary which is not a data type
that types and providers in general know how to deal with.

This is now changes such that the DeferredResolver replaces
all such instances with an ASCII-8BIT encoded string as this
is backwards compatible with how such binary strings were handled
in earlier versions when serializing binary content with PSON.

This implementation ensures symmetrical serialization/deserialization
while allowing Binary to become ASCII-8BIT string in the post processing
stage.

Only small issue is that the DeferredResolver is no longer a great name
for the resolver.