(PUP-9133) Make agent side loader load all modules by using wildcard
This introduces the abolity to give a whilecard as "module name" for
a module loader. This will have the same effect as setting `nil`
(which indicates that it is a kind of global loader), but in contrast
it will accept any namespace and attempt to load from it where a `nil`
module name means *only* accepting unqualified name (top/global
namespace).