(PUP-8987) Add resource_api to gemspec/Gemfile
The resource API will be shipping as part of puppet-agent 6.0, and we
are currently planning to encourage the community and partners to
build content for it. It should be part of the default Puppet Gem
distribution so that it can be picked up by PDK and other
non-package-based workflows easily.