(PUP-9094) Switch rest client back to our HttpPool
We had been working to have our rest client use the httpclient gem for
cert related HTTP requests. Unfortunately, we weren't able to finish
that work in time for Puppet 6.0.0. This commit switches those requests
back to using Puppet::Network::HttpPool.

Unfortunately, this isn't a clean switch back. We are preserving all the
existing work as much as possible so we can more easily get back to the
httpclient gem in the future. For now, we have to sacrifice some nice
new functionality in favor of the old way of doing things. Notably, if
users have an invalid or out of date cert on their agent, they have to
remove it as the automation will not update it. This is consistent with
how things used to work prior to the httpclient gem.