(PUP-9135) Treat task metadata referring to missing module as invalid
Previously, we were raising a ModuleNotFound error in the case where a
task depended on a file from another module that didn't exist. Raising
that error caused a failure that was difficult to debug (the task was
treated as simply not existing). In reality, this scenario is generally
a problem with the metadata just like any other metadata error. Because
of that, we now raise InvalidMetadata, which causes the actual error
message to propagate through puppet-server and orchestrator, to the
client.

Additionally, this error message was vague and fairly unhelpful as it
didn't explain *why* it was trying to load that module in the first
place. That message has been improved.