(PUP-9135) Check for path traversal before checking file existence
Previously, we were checking whether the referred file existed *before*
checking whether it was a path traversal. This allowed a potential
disclosure of file existence, through the roundabout method of deploying
some task metadata using a path traversal to refer to a path of interest
and then attempting to run the task and checking the error message. We
now check for path traversal *first*, which also avoids checking it for
every separate file in a directory.