(PUP-9137) Avoid reloading a 3x function
Before this, when a 4x loader tries to load a 3x function it
would not check if the 3c function had already been loaded via
the 3x API. This would cause the function to be reloaded and
the information already loaded would be overwritten and this causes
a warning to be issued.

Now, a check is made if function is already loaded and if so using
the loaded function instead of loading it again.