(PUP-9106) Ensure SYSTEM ACE entries are Full Control
Previously it was possible to specify an octal mode for setting file permissions
which would end up with the SYSTEM account in Windows not having Full Control.
This could then cause permission issues as the SYSTEM account is a priveledged
account.  This commit intercepts attempts to set the mode on files/directories
for the SYSTEM account to something other than Full Control; and then issues a
warning and corrects the ACE back to Full Control.