(PUP-5475) Add default timeout for waiting on pending windows services
Since we added waits on pending statuses for windows we uncovored an issue:
if the service does not supply a wait hint puppet will not wait at all for the
service to move out of pending. Since we can't assume everything supplies a
wait hint (puppet, for example, does not) this commit updates the wait
functionality in puppet to include a floor of at least 30 seconds: if the
service provides a wait hint less than 30 seconds or none at all puppet will
wait at least that long.