(PUP-5475) Update windows service acceptance tests with new timeout
We added a floor on the timeout when puppet is waiting for pending states to
30 seconds. This broke the service acceptance tests because they were set to
'break' at 4 seconds (which is impossible with the new 30 second minimum
timeout). This commit updates the tests to break at 30 seconds by having the
broken services wait for 60 seconds.

This wait time is long, and will add just about a minute to the test run for
services. However we believe it's important to test the failure state that
puppet will actually time out, so we believe it's worth the extra time.