(PUP-9117) Fix exec/should_run_command_as_user test failure on Arista
This test failed in the initial agent.user_absent(username) step,
which is meant to ensure that our stubbed user does not exist on
our agent before we run the test. The reason it failed there is
because Beaker::Host's user_absent (and user_present) functions do
not work correctly on Arista (EOS). This is because `getent passwd <user>`
always returns an exit code of 0 on Arista, even if the user does
not exist. We use that command to check for user existence/nonexistence
before ensuring that they're absent/present, respectively.

This commit fixes the test by creating wrappers around `user_present`
and `user_absent` until the issue is fixed in Beaker.