(PUP-9144) Add the ability to continue building a catalog with PAL
Before this, after the initial compilation (controlled by manifest/code
options) with PAL for catalog compilation there was no way to get lazy
constructs evaluated or to again run validation. Lazy constructs and
validation was performed for the initial compilation so the problem
addressed here is not super critical, but makes it impossible to
continue to build up the catalog with dynamic sequences.

To fix this, the PAL API for the exposed `CatalogCompiler` class is
extended with three methods: `compile_additions`, `evaluate_additions`
and `validate`, where the first method calls the other two. This allows
a dynamic sequence of operations that for example uses `evaluate_string`
to add definitions, declare a user defined resource in the catalog and
make it (and all other queued lazy constructs) to be run on demand.

This alters the standard compiler so that it exposes the relevant
methods to its subclasses by making some methods `protected`.