(PUP-9155) Make `puppet cert` error when args are passed
In Puppet 6 we replaced `puppet cert` with `puppetserver ca`. We tried
to make `puppet cert` error when it was used, but accidentally only
caused the error message to print when no subcommands or args were
passed. This commit updates `puppet cert` to error helpfully in all cases.
It also causes it to always exit 1.