(PUP-3647) Add ability to prune resource state cache & TTL setting
The new `statettl` setting sets how long to store information about when a
resource was last checked in the `statefile`. Previously, the `statefile`
would grow unbounded, which could lead to quite lengthy agent run times
due to loading & saving the file. We now default to saving only 7 days
worth of information in the `statefile`. Using a value of `0`, or
`unlimited` will restore the previous behavior of never removing entries
from the file.