(PUP-9111) Retry reading overrides plist to avoid launchd race
When restarting a service, the launchd provider calls `stop` and then `start`.
The `stop` method executes `launchctl unload -w <plist path>`, and then calls
`enable`. The `enable` method reads the disable overrides pilst and modifies it.

However, there is a race. While the `launchd` process is changing states, the
disables plist is sometimes zero-length. When this occurs, `read_plist` fails,
causing the refresh to fail.

Add a `read_overrides` method which retries the read_plist with a short sleep.