(PUP-6716) Purge local users on Windows
Attempting to purge local users was failing on Windows, since the code
for that was Unix-centric, assuming a numeric user ID. Windows uses
a string for this (the SID), so a comparison was failing.

This adds in some logic to check for system users by marking some SID
formats as system users, via regex.