(MAINT) Fix json_spec for binary files
PR #7063 introduced a change to serialize binary in ASCII; tests in
the json_spec with binary data started failing after merge. This change
modifies the test to provide the need context for a loader and updates
the expects to look use the `binary_buffer` method on the content of a
resource in the catalog.