(PUP-6542) Refactor members property to be more like the groups property
The groups property in the User resource is derived from Puppet::Property::List.
This class also manages inclusive vs. minimum membership so that Puppet
always reports the right change notifications. The members property in
the Group resource mirrors the groups property conceptually in that they
both take in the same input and that they both implement the concept of
inclusive vs. minimum membership. However for some reason, the members
property is implemented differently than the groups property. This
implementation is bad, because it leaves the management of inclusive vs.
minimum membership to the provider when it shouldn't: inclusive vs. minimum
membership can and should be managed at the property-level. Leaving it
to the provider causes Puppet to report the wrong change notification
because it gets this information from the property.

This commit refactors the members property to be consistent with the
groups property of the User resource. It essentially copy-pastes the
groups property code over, with some slight differences:

* The members property can be specified as a comma-separated list of
members. We still need to preserve this behavior for backwards
compatibility.

* Likewise, we need to modify the inclusive? function to use the
auth_membership parameter directly since the latter's a Boolean on
the Group resource instead of an enum. We should probably change this
at some point.

It may be worthwhile to extract the groups and members properties into a
common class or module at some point in the future. Since the copy-pasted
code is relatively small, and since it is not exactly obvious what this
common class or module should be called, I have chosen not to do this
here to keep things simple.

This change breaks our existing group providers, which we will fix in
subsequent commits. Specifically, the providers must be modified to:
* Return a comma-separated list instead of an array in the property
getter
* Accept a comma-separated list instead of an array in the property
setter