(maint) Remove dead caching code from configurer
Attempting to read from the cached `:configuration` section to get the
`:compile_time` would create an empty `:configuration` section that would
get written back to the state.yaml. The `:compile_time` is never written
to in the cache, and the attr_reader (and instance variable) are never
used beyond the initialization, so we can safely delete the code entirely.