(PUP-9163) Fix lexer problem with heredoc and sublocator
Before this, heredoc set up a locator intended to define a subspace
for the heredoc string (for lexing its content to support
interpolation). What it ended up doing was to create a subspace offset
not from the total source string, but from within the heredoc string.
This resulted in offsets being calculated to point to char positions in
an undefined manner (depending of length of heredoc, its position and
what follows). This would then result in error messages pointing to the
wrong place in the source code as well as wrongly computing if a `[` is
an array start or not by checking if the preceding char is a space since
it would end up looking at some undefined position.

This was made worse (or the same) because the lexer would use the
returned char_offset as an offset in its subspace string instead of in
the total space.