(PUP-5216) Configure the attributes property for Windows users
This commit configures the attributes property to work with
Windows users via. the following steps:

* It decouples the property from AIX by requiring providers to
have the manages_attributes feature instead of the manages_aix_lam
feature in order to use it. We still need the manages_aix_lam feature,
however, as other parameters/properties could still rely on it, like
the ia_load_module parameter in the User resource.

* It introduces the implements_attribute_membership feature to
separate providers that use the attributes property AND also
implement inclusive/minimum attribute membership from providers that
just use the attributes property, but do not implement attribute
membership. The latter is required to make the attributes property
work with Windows users since the concept of attribute membership
does not really make sense there.