(PUP-9189) Refactor the state transition code in Windows::Service
Previously, the transition_service_state helper used by the
Windows::Service.start and Windows::Service.stop methods had
a verbose interface that made it a bit more confusing to
understand what was going on than it needed to be. For example,
it was not clear what the difference between the preceding_transition
and preceding_state parameters is from the parameter names alone.

This commit refactors the transition_service_state helper to more
consistently model the Windows service state transition diagram. It
also extends the helper to support the possibility that the service
may have already transitioned to the final state, whereby it will
return early (without performing the transition), or that it is is in
the middle of a pending transition to the final state, whereby it
will wait for the pending transition to finish, and then return early
(without performing the transition). This lets Windows::Service.start
and Windows::Service.stop noop if the service is already starting/started
or stopping/stopped, which is consistent with e.g. systemd.