(PUP-9180) Convert nested :undef to nil in 3x converter
Before this, nested :undef values were converted to the given
"undef value" (either nil, or empty string). When this was changed for
puppet 6.0.0 the language no longer produces :undef and the logic was
simplified. This did however break tests that assumed that functions
would be called with :undef.

This commit makes one change; a nested `:undef` will be converted to
`nil`.