(maint) Don't care about subseconds in state file pruning tests
JRuby can't reliably round-trip Time objects keeping the subseconds the
same. Fortunately, we don't actually care about subsecond resolution for
the pruning tests, so we can round off the subseconds to 0, to try to get
JRuby to consistently round-trip the Time.

This also bumps the amount the stale cache entry has been stale for to
give a buffer around possible rounding of the time.