(PUP-2997) Only handle_logdest_arg if given an arg and we haven't already set the destination
The plan is to use handle_logdest_arg to also handle setting the log
destination if logdest was set via a config file. Since a number of the
applications either override #setup, or #setup_logs, we'll need to start
calling handle_logdest_arg from more places. Rather than duplicating the
nil and "did we already do this" checks everywhere, we can do those checks
as a guard in the method itself.