(PUP-2997) Allow setting --logdest via puppet.conf
The --logdest option can now be set via the logdest setting in
puppet.conf.

Unfortunately, handling setting the logging destination itself via an
:on_initialize_and_write hook ended up playing very poorly with a large
number of tests that were expecting to be able to arbitrarily reinitialize
Puppet's settings without having it affect the logging setup so they could
look for warnings/deprecations/errors.

Moving the setting of the log destination itself into the applications
turned out to be less problematic, even though it's a bit awkward to work
with because of some of the applications overriding #setup and #setup_logs.