(PUP-8971) Fix unit test of Encrypted for Puppet 6
Before this, the unit tests for Encrypted would not work since
Puppet 6.0 got rid of the Ruby based CA. Now the tests use
the TestCa in combination with mocking. The mocking is needed
since there is no storage of certificates configured and the logic
expects to be able to create a Puppet::SSL::Host and obtain certificates
and private keys for hosts.