(PUP-9193) Make puppet honor installation exit code, still suppress console
Previously, the windows package provider for exe packages would ignore the
exit code from the installation because cmd.exe would return the exit code
from the cmd.exe executable and not the installer executable.

This commit removes the use of cmd.exe to execute install/uninstall commands.
We now ensure that processes don't open a new window by using the
CREATE_NO_WINDOW flag passed to Process.create (which in turn will send that
flag to CreateProcessW in the windows API).

We specify to pass the CREATE_NO_WINDOW flag with the :suppress_window option
added to Puppet::Util::Execution.execute

This should solve all our problems: the process will wait for execution to
finish, not open a new console, and exit with the correct exit code