(PUP-9156) Add action to clean an agent's client cert and key
Add a 'clean' action to 'puppet ssl' to delete the client cert, private key,
public key, certificate request, and private key passfile for the current host
only.

When run on the CA host, the action will not revoke or clean certs issued by
the CA. To do that, use the `puppetserver ca clean` command.