(maint) Update dev dep of puppetserver-ca to 1.1
Previously we pulled in a pre-1.0 version of puppetserver-ca as a
development dependency. With Puppet Server 6.0.0 we shipped 1.0.0 and
consequently developers should have a 1+ version available during
development.

This updates to the latest y release of puppetserver-ca and ensures the
bundle installs should receive the latest y release of the 1 major
series.