(PUP-9156) Protect against cleaning the CA's client cert
Protect against running `puppet ssl clean` on the CA, as it will delete
the client cert and key that the `puppetserver ca` CLI needs to
communicate with the CA REST API.

If the agent is an agent only or is on the puppetserver host, but it is
not the CA, then allow `puppet ssl clean` to behave as it previously.

If the agent is on the CA host, then ensure the cert we're about to
clean has already been removed from the CA. If it has, then clean the
agent's local copy. If it hasn't, print an error and exit.