(PUP-9156) Use Puppet's console logger
Use Puppet's console logger so error warnings and errors are colorized. Also add
support for verbose and debug CLI options so it's possible to see debug messages
from puppet internals. Update usage text.

The `to_output` matcher can not intercept stream output that uses a reference
to $stdout/$stderr that was stored before the matcher was used, which is the
case with puppet's console logger. So use our `have_printed` matcher instead.

If we submit a request, and the cert isn't immediately signed, then emit an
"info" message and return success, as will be the case when autosigning is not
enabled. If we explicitly try to download a cert, and it isn't signed yet, emit
an error and exit with 1.