(maint) Refactor ::get_latest_version
The method needed two arguments: the origin of the package we wanted to check
for updates, and the list of packages with updates.

Now, only pass the package origin, and let the get_latest_version run `pkg
version` if the list of packages with updates is not available.  Because the
list can become outdated, provide two methods to get the list of packages with
updates: ::get_version_list and ::cached_version_list.  The first one forces
the refreshing of the list if it is already available, the second only get the
list if it is not already available.