(maint) Change osx user gid check
Beaker 4 changes how we get the user gid. In earlier versions of beaker,
we run `dscacheutil -q user -a name #{name}`, but in beaker 4 we switch
to `id -P #{name}`. Since this test is currently run with beaker 3, and
we're getting ready to switch it to beaker 4, we need this test to be
able to handle both methods of parsing the user gid.