(PUP-9317) Use new resource initialization methods when realizing
This backports the fix for PUP-8002 to 5.5.x.

Prior to this commit, the resource collection code used a pre-pops
API for finding resource types. This triggered the same error as
exported/collected resources ran into in PDB-3734.

This patch updates the code to use the new pops-related API for
finding resources, similar to the fix for the puppetdb terminus,
but without the conditions excluding classes and nodes since they
could be valid in this context.